refactor(wallet): simplify architecture and introduce sqlite backend
Simplify the verbosity of the previous architecture by skipping the
old App actor altogether and use directly a sync actor that handles
jsonrpc requests in each of its threads.

Also, introduce the use of sqlite as the storage solution which will
greatly simplify the amount of work on the storage side.