fix(data_structures): add contains method and ignore commit/reveal hash
If there are two commits with equal pkh for the same data request, we
only accept the first one and ignore the rest as duplicates, even if
they are different. Same with reveals.

This fixes an issue where one commit can overwrite another and we enter
an infinite loop since the received commit is always considered new.