feat(config): add phase-1 nodes as known peers
Multiple addresses in the same network have been avoided just as
a way to verify the peer discovery mechanism, the "feeler" and the
bucketing system.

In other words: if one of your IPs is not here, it doesn't mean
your node will not receive inbound connections — the network as a
whole should be now able to discover them in a matter of minutes.