test(data_structures): update priority tests
We need to update the tests that use priorities_factory because the inclusion of
st (stake) and ut (unstake) priorities introduces additional random values into the
priority generation process. This change affects the sequence of random numbers
generated by the pseudorandom number generator (prng) and, consequently, the resulting Priorities objects.